+++ src/java/GUI/Bean/CreateAccount.java
@@ -37,7 +37,7 @@ public class CreateAccount {
 
     private String email;
 
-    private int telephone;
+    private long telephone;
 
     private String addres;
 
@@ -110,11 +110,11 @@ public class CreateAccount {
         this.email = email;
     }
 
-    public int getTelephone() {
+    public long getTelephone() {
         return telephone;
     }
 
-    public void setTelephone(int telephone) {
+    public void setTelephone(long telephone) {
         this.telephone = telephone;
     }
